Added entries saving feature
Added a feature in which the entries that were webscraped from the
Acumen webpage are stored in a csv file. The purpose of this csv file is
to improve performance by reducing the amount of entries we need to
extract from the website (we don't need to extract all entries everytime
we run the script).